refactor: rename packages and remove empty lines
