+++ public/fonts/fonts.css
@@ -1,13 +1,11 @@
-
-
-@font-face{
-    font-family:'ALS Sirius';
-    src:url('/fonts/ALS-Sirius/ALS_Sirius_Regular_0.95.otf') format('opentype');
-    font-weight:400;
+@font-face {
+  font-family: "ALS Sirius";
+  src: url("/fonts/ALS-Sirius/ALS_Sirius_Regular_0.95.otf") format("opentype");
+  font-weight: 400;
 }
 
-@font-face{
-    font-family:"ALS Sirius";
-    src: url('/fonts/ALS-Sirius/ALS_Sirius_Bold_0.95.otf') format(opentype);
-    font-weight:700;
+@font-face {
+  font-family: "ALS Sirius";
+  src: url("/fonts/ALS-Sirius/ALS_Sirius_Bold_0.95.otf") format(opentype);
+  font-weight: 700;
 }
