--- src/assets/svg/Cross1/cross1.module.css
@@ -1,7 +1,0 @@
-
-.cross{
-    position:absolute;
-    top:16px;
-    right:16px;
-    cursor:pointer;
-}
