+++ src/assets/svg/CrossSvg/CrossSvg.jsx
