+++ src/assets/svg/CrossSvg/cross1.module.css
@@ -1,0 +1,6 @@
+.cross {
+  position: absolute;
+  top: 16px;
+  right: 16px;
+  cursor: pointer;
+}
