--- src/assets/svg/Heart/heart.module.css
@@ -1,8 +1,0 @@
-
-
-.heart{
-    position:absolute;
-    right:5px;
-    top:5px;
-    cursor:pointer
-}
