+++ src/assets/svg/HeartSvg/heart.module.css
@@ -1,0 +1,6 @@
+.heart {
+  position: absolute;
+  right: 5px;
+  top: 5px;
+  cursor: pointer;
+}
