+++ src/сomponents/Book/book.module.css
@@ -1,36 +1,35 @@
-
-.book-container{
-    width:160px;
-    height:350px;
+.book-container {
+  width: 160px;
+  height: 350px;
 }
 
-.image-container{
-    display:flex;
-    justify-content:center;
-    align-items:center;
-    width:100%;
-    height:250px;
-    position:relative;
-    margin-bottom:10px;
+.image-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 250px;
+  position: relative;
+  margin-bottom: 10px;
 }
 
-img{
-    object-fit:contain;
-    max-width: 100%;
-    max-height:100%;
-    border-radius:10px;
-    cursor:pointer;
+img {
+  object-fit: contain;
+  max-width: 100%;
+  max-height: 100%;
+  border-radius: 10px;
+  cursor: pointer;
 }
 
-.bookName{
-    font-weight: 600;
-    cursor:pointer
+.bookName {
+  font-weight: 600;
+  cursor: pointer;
 }
 
-.authorName{
-    color:#5a617c
+.authorName {
+  color: #5a617c;
 }
 
-.genre{
-    color:#22252f
+.genre {
+  color: #22252f;
 }
