+++ src/сomponents/Footer/footer.module.css
@@ -1,5 +1,4 @@
-
-footer{
-    height:100px;
-    background-color:var(--main-color)
+footer {
+  height: 100px;
+  background-color: var(--main-color);
 }
