+++ src/сomponents/MenuUser/menuUser.module.css
@@ -8,6 +8,7 @@
   flex-direction: column;
   align-items: center;
   border-left: 5px solid var(--border-color);
+  z-index: 1001;
 }
 
 .menu-user-container {
