+++ src/сomponents/Sorting/sorting.module.css
@@ -1,44 +1,43 @@
-
-.sorting-wrapper{
-    position:relative
+.sorting-wrapper {
+  position: relative;
 }
 
-.sorting-container{
-    position:absolute;
-    top:120%;
-    right:-50%;
-    transform:translateX(30%);
-    z-index:999;
-    border:2px solid black;
-    background-color:white;
-    width:130px;
-    border-radius:10px;
+.sorting-container {
+  position: absolute;
+  top: 120%;
+  right: -50%;
+  transform: translateX(30%);
+  z-index: 999;
+  border: 2px solid black;
+  background-color: white;
+  width: 130px;
+  border-radius: 10px;
 }
 
-.title{
-    font-size:1.2em;
-    text-align:center;
-    margin-bottom:8px;
-    user-select:none;
+.title {
+  font-size: 1.2em;
+  text-align: center;
+  margin-bottom: 8px;
+  user-select: none;
 }
 
-.radioLabelContainer{
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    margin-bottom: 2px;
+.radioLabelContainer {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  margin-bottom: 2px;
 }
 
-.input-radio{
-    width: 15px; 
-    height: 15px; 
-    margin-left: 5px;
-    cursor:pointer;
+.input-radio {
+  width: 15px;
+  height: 15px;
+  margin-left: 5px;
+  cursor: pointer;
 }
 
-@media(max-width:576px){
-    .sorting-container{
-        transform:none;
-        right:-30%
-    }
+@media (max-width: 576px) {
+  .sorting-container {
+    transform: none;
+    right: -30%;
+  }
 }
